% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{build_m2efm}
\alias{build_m2efm}
\title{Build the M2EFM model.}
\usage{
build_m2efm(data, outcome, clin, covariates, standardize = FALSE, alpha = 0)
}
\arguments{
\item{data}{data.frame with molecular data.}

\item{outcome}{Survival object from survival package, or binomial outcome}

\item{clin}{data.frame with clinical data}

\item{covariates}{vector of column names in clinical data}

\item{standardize}{whether or not to standardize the data}
}
\value{
a vector of genes to use for models
}
\description{
Build the M2EFM model.
}

