% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{evaluate}
\alias{evaluate}
\title{Calculates prognostic statistics based on m2eQTLs.}
\usage{
evaluate(meth = NULL, exp = NULL, clin = NULL, eqtls = NULL,
  gene_type = NULL, covariates = c(), gene_sig = NULL, gene_seed = NULL,
  genes = NULL, num_genes = NULL, num_probes = NULL,
  intersect_data = FALSE, integrate_data = FALSE, single_data = "exp",
  time_to_event_col = NULL, event_col = NULL, rescale = TRUE,
  monte_carlo_size = 100, training_prop = 0.7, pam50risk = NULL,
  val = list(), val_clin = list(), val_pam50 = list(), alpha = 0)
}
\arguments{
\item{meth}{a MethylationData object, or a data.frame with DNA methylation Beta-values}

\item{exp}{an ExpressionData object, or a data.frame with gene expression values}

\item{clin}{a data.frame of clinical data for both meth and exp samples}

\item{eqtls}{object returned from get_m2eqtls()}

\item{gene_type}{type of gene list being provided ("trans", "cis", "random", "pre") where trans and cis are those types of m2eQTLs and random is a random set of genes}

\item{covariates}{a vector of covariates to include from the clinical data}

\item{gene_sig}{file to store the gene list in (or load it from)}

\item{gene_seed}{random seed to generate random gene list from}

\item{num_genes}{number of genes to use for a random gene list}

\item{num_probes}{number of probes to use for a random probe list}

\item{intersect_data}{whether to intersect samples for meth and exp (automatically done if integrate_data=TRUE)}

\item{integrate_data}{whether or not to integrate meth and exp values into a single data frame for the model}

\item{single_data}{which single data type to use if integrate_data=FALSE, can be "exp" or "meth"}

\item{time_to_event_col}{name of column with time to event}

\item{event_col}{name of column with sample event status, or outcome for binomial models}

\item{monte_carlo_size}{how many random splits of the training data to perform}

\item{training_prop}{proportion of the data to use for training}

\item{val}{a list of validation data sets (list of data.frame objects, with combined methylation and expression values or just one or the other)}

\item{val_clin}{a list of validation clinical data, parallel to val}

\item{alpha}{a value in [0,1] where 0 is Ridge penalty and 1 is LASSO}
}
\value{
a list of results
}
\description{
Calculates prognostic statistics based on m2eQTLs.
}

