% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionData.R
\docType{data}
\name{ExpressionData}
\alias{ExpressionData}
\title{ExpressionData}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExpressionData
}
\description{
A class for holding expression profile data.
}
\keyword{datasets}

