% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2eQTL.R
\name{get_entrez_ids}
\alias{get_entrez_ids}
\title{Takes a vector of gene symbols and converts them to entrez ids.}
\usage{
get_entrez_ids(gene_symbols)
}
\arguments{
\item{gene_symboles}{a vector of gene symbols}
}
\value{
a data.frame of gene symbols and corresponding entrez ids
}
\description{
Takes a vector of gene symbols and converts them to entrez ids.
}

