% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2eQTL.R
\name{get_gene_locs}
\alias{get_gene_locs}
\title{Takes a data.frame with entrez ids and gene symbols and builds a table of 
gene locations.}
\usage{
get_gene_locs(entrez_ids)
}
\arguments{
\item{entrez_ids}{data.frame with first column gene symbols and second column the corresponding entrez ids}
}
\value{
a data.frame with gene locations
}
\description{
Takes a data.frame with entrez ids and gene symbols and builds a table of 
gene locations.
}

