% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{get_genes}
\alias{get_genes}
\title{Returns the genes to use from m2eQTL analysis.}
\usage{
get_genes(eqtls, gene_type, integrate_data = FALSE)
}
\arguments{
\item{eqtls}{object returns from get_m2eqtls()}

\item{gene_type}{"cis" or "trans"}
}
\value{
a vector of genes to use for models
}
\description{
Returns the genes to use from m2eQTL analysis.
}

