% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2eQTL.R
\name{get_m2eqtls}
\alias{get_m2eqtls}
\title{Finds m2eQTLs from gene expression and DNA methylation data.}
\usage{
get_m2eqtls(probe_list_loc = NULL, meth_data = NULL, exp_data = NULL,
  gene_list = NULL, num_cis = 1e+10, num_trans = 100, num_probes = 300)
}
\arguments{
\item{probe_list_loc}{path to a list of probe ids, used to filter methylation data}

\item{meth_data}{path to a matrix of methylation data or a MethylationData object}

\item{exp_data}{path to expression data or an ExpressionData object}

\item{gene_list}{a file with a list of genes to filter to (optional)}

\item{num_cis}{number of cis m2eQTLs to return}

\item{num_trans}{number of trans m2eQTLs to return}

\item{num_probes}{number of top probes to use in analysis}
}
\description{
Finds m2eQTLs from gene expression and DNA methylation data.
}

