% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{get_pam50}
\alias{get_pam50}
\title{Get pam50 rorS scores for sample with gene expression data. It
should go without saying that these should be breast cancer samples.}
\usage{
get_pam50(exp)
}
\arguments{
\item{exp}{a data.frame with gene expression values}
}
\value{
a list with subtype calls and rorS scores for the samples in exp
}
\description{
Get pam50 rorS scores for sample with gene expression data. It
should go without saying that these should be breast cancer samples.
}

