% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2eQTL.R
\name{get_probe_locs}
\alias{get_probe_locs}
\title{Takes a vector of methylation probe ids and builds a table of genomic 
probe loci.}
\usage{
get_probe_locs(probe_list)
}
\arguments{
\item{probe_list}{vector of probe ids}
}
\description{
Takes a vector of methylation probe ids and builds a table of genomic 
probe loci.
}

