% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{get_probes}
\alias{get_probes}
\title{Returns the probes to use from m2eQTL analysis.}
\usage{
get_probes(eqtls)
}
\arguments{
\item{eqtls}{object returns from get_m2eqtls()}
}
\value{
a vector of probes to use for models
}
\description{
Returns the probes to use from m2eQTL analysis.
}

