% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{get_random_genes}
\alias{get_random_genes}
\title{Returns the genes to use selected at random.}
\usage{
get_random_genes(exp, seed, num_genes)
}
\arguments{
\item{exp}{data.frame of gene expression data}

\item{seed}{a random seed}

\item{num_genes}{the number of genes to take at random}
}
\value{
a vector of genes to use for models
}
\description{
Returns the genes to use selected at random.
}

