% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2eQTL.R
\name{load_exp}
\alias{load_exp}
\title{Takes the path to a gene expression matrix and return an ExpressionData object.}
\usage{
load_exp(exp_data_loc)
}
\arguments{
\item{exp_data_loc}{path to a matrix of gene expression values}
}
\value{
a ExpressionData object
}
\description{
Takes the path to a gene expression matrix and return an ExpressionData object.
}

