% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2eQTL.R
\name{load_meth}
\alias{load_meth}
\title{Takes the path to a DNA methylation matrix and return a MethylationData object.}
\usage{
load_meth(meth_data_loc)
}
\arguments{
\item{meth_data_loc}{path to a matrix of DNA methylation beta values}
}
\value{
a MethylationData object
}
\description{
Takes the path to a DNA methylation matrix and return a MethylationData object.
}

