% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationData.R
\docType{data}
\name{MethylationData}
\alias{MethylationData}
\title{MethylationData}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MethylationData
}
\description{
A class for holding methylation profile data.
}
\keyword{datasets}

