% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationData450k.R
\docType{data}
\name{MethylationData450k}
\alias{MethylationData450k}
\title{MethylationData450k}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MethylationData450k
}
\description{
A class for holding methylation profile data from Illumina 450k.
}
\keyword{datasets}

