% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{mm}
\alias{mm}
\title{Display a message to the user.}
\usage{
mm(text, appendLF = TRUE)
}
\arguments{
\item{text}{text of the message to display}
}
\value{
nothing
}
\description{
Display a message to the user.
}

