% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{plot.M2EFM_evaluation}
\alias{plot.M2EFM_evaluation}
\title{Create boxplots of the evaluation results.}
\usage{
plot.M2EFM_evaluation(eval_result, pal = NULL, ...)
}
\arguments{
\item{eval_result}{object returned by evaluate()}
}
\value{
nothing
}
\description{
Create boxplots of the evaluation results.
}

