% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{predict.M2EFM_models}
\alias{predict.M2EFM_models}
\title{Predict the M2EFM model on new data.}
\usage{
\method{predict}{M2EFM_models}(fit, data, clin, covariates)
}
\arguments{
\item{fit}{the M2EFM models to predict - the result of build_m2efm}

\item{clin}{data.frame with clinical data}

\item{covariates}{vector of column names in clinical data}

\item{new_data}{data.frame with molecular data.}
}
\value{
a data.frame of up to two columns of predictions: for molecular and covariates and molecular only
}
\description{
Predict the M2EFM model on new data.
}

