% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileData.R
\docType{data}
\name{ProfileData}
\alias{ProfileData}
\title{ProfileData}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProfileData
}
\description{
A class for holding gene or methylation profile data.
}
\section{Fields}{

\describe{
\item{\code{values}}{The data values.}
}}
\keyword{datasets}

