% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{reduce_stage}
\alias{reduce_stage}
\title{Converts AJCC stages to a simple four stage form.}
\usage{
reduce_stage(x)
}
\arguments{
\item{x}{a vector of AJCC stages (e.g. Stage IA)}
}
\value{
a vector of simplified AJCC stages
}
\description{
Converts AJCC stages to a simple four stage form.
}

