% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{scale_samples}
\alias{scale_samples}
\title{Scales individual samples to another dataset.}
\usage{
scale_samples(profile, means, ranges)
}
\arguments{
\item{profile}{a data.frame with gene expression values}

\item{means}{a vector of means by gene from original dataset}

\item{ranges}{a vector of ranges or sds by gene from original dataset}
}
\value{
a data.frame of gene expression values with each gene scaled
}
\description{
Scales individual samples to another dataset.
}

