% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{scale_values}
\alias{scale_values}
\title{Center and scale the expression values.}
\usage{
scale_values(profile)
}
\arguments{
\item{profile}{a data.frame with gene expression values}
}
\value{
a data.frame of gene expression values with each gene in [-1,1]
}
\description{
Center and scale the expression values.
}

