% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{summary.M2EFM_evaluation}
\alias{summary.M2EFM_evaluation}
\title{Summarize the results of an M2EFM evaluation.}
\usage{
\method{summary}{M2EFM_evaluation}(eval_result, nph = FALSE, ...)
}
\arguments{
\item{eval_result}{object returned by evaluate()}
}
\value{
nothing
}
\description{
Summarize the results of an M2EFM evaluation.
}

