% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2efm.R
\name{summary.M2EFM_models}
\alias{summary.M2EFM_models}
\title{Summarize the results of an M2EFM models.}
\usage{
\method{summary}{M2EFM_models}(m2efm_result, ...)
}
\arguments{
\item{m2efm_result}{object returned by build_m2efm()}
}
\value{
nothing
}
\description{
Summarize the results of an M2EFM models.
}

